+++ lua/windows.lua
@@ -72,6 +72,7 @@ workspace "ECM-R"
         targetname "%{cfg.architecture}/ecm-r.%{cfg.architecture}"
 		language "c++"
 		cppdialect "c++17"
+		buildoptions { "/utf-8" }
 		kind "sharedlib"
 		warnings "off"
 
@@ -126,6 +127,7 @@ workspace "ECM-R"
 			"../src/app/input/**",
 			"../src/app/audio/**",
 			"../src/app/settings/**",
+			"../src/app/localization/**",
 
 			"../src/utils/fs/**",
 			"../src/utils/logger/**",
@@ -136,12 +138,24 @@ workspace "ECM-R"
 
        filter "platforms:Win-x86"
 			postbuildcommands {
+				"IF NOT EXIST \"$(OutDir)x86\" mkdir \"$(OutDir)x86\"",
+				"IF NOT EXIST \"$(OutDir)x86\\ecm-r\\translations\" mkdir \"$(OutDir)x86\\ecm-r\\translations\"",
+				"copy /y \"$(ProjectDir)..\\ecm-r\\translations\\en.ini\" \"$(OutDir)x86\\ecm-r\\translations\\en.ini\"",
+				"copy /y \"$(ProjectDir)..\\ecm-r\\translations\\es.ini\" \"$(OutDir)x86\\ecm-r\\translations\\es.ini\"",
 				"IF NOT EXIST \"$(OutDir)\"x86 mkdir \"$(OutDir)x86\\",
                   "copy /y \"$(OutDir)x86\\ecm-r.x86.dll\" \"$(OutDir)x86\\ecm-r.x86.asi\"",
 			}
 
        filter "platforms:Win-x64"
 			postbuildcommands {
+				"IF NOT EXIST \"$(OutDir)x86\" mkdir \"$(OutDir)x86\"",
+				"IF NOT EXIST \"$(OutDir)x86_64\" mkdir \"$(OutDir)x86_64\"",
+				"IF NOT EXIST \"$(OutDir)x86\\ecm-r\\translations\" mkdir \"$(OutDir)x86\\ecm-r\\translations\"",
+				"copy /y \"$(ProjectDir)..\\ecm-r\\translations\\en.ini\" \"$(OutDir)x86\\ecm-r\\translations\\en.ini\"",
+				"copy /y \"$(ProjectDir)..\\ecm-r\\translations\\es.ini\" \"$(OutDir)x86\\ecm-r\\translations\\es.ini\"",
+				"IF NOT EXIST \"$(OutDir)x86_64\\ecm-r\\translations\" mkdir \"$(OutDir)x86_64\\ecm-r\\translations\"",
+				"copy /y \"$(ProjectDir)..\\ecm-r\\translations\\en.ini\" \"$(OutDir)x86_64\\ecm-r\\translations\\en.ini\"",
+				"copy /y \"$(ProjectDir)..\\ecm-r\\translations\\es.ini\" \"$(OutDir)x86_64\\ecm-r\\translations\\es.ini\"",
 				"IF NOT EXIST \"$(OutDir)\"x86 mkdir \"$(OutDir)x86\\",
 				"IF NOT EXIST \"$(OutDir)\"x86_64 mkdir \"$(OutDir)x86_64\\",
               "copy /y \"$(OutDir)x86\\ecm-r.x86.dll\" \"$(OutDir)x86\\ecm-r.x86.asi\"",